+++ src/functions/on-call/update-paid/index.ts
@@ -1,0 +1,16 @@
+import { handlerPath } from "@libs/handler-resolver";
+
+export default {
+  handler: `${handlerPath(__dirname)}/handler.main`,
+  events: [
+    {
+      httpApi: {
+        method: 'post',
+        path: '/update-paid',
+        authorizer: {
+          name: "timebankKeycloakAuthorizer"
+        }
+      },
+    },
+  ],
+}
