+++ src/functions/on-call/update-paid/mockData.json
@@ -2,5 +2,5 @@
     "headers": {
       "Content-Type": "application/json"
     },
-    "body": "{}"
+    "body": "{\"year\": 2019, \"week\": 1, \"paid\": true}"
 }
