--- src/functions/update-paid/index.ts
@@ -1,6 +1,0 @@
-import { handlerPath } from "@libs/handler-resolver";
-
-export default {
-  handler: `${handlerPath(__dirname)}/handler.main`,
-  events: [  ],
-}
